TCA-1279 - hide header & footer for hiring manager view in modal mode
