+++ src-ts/tools/learn/learn-lib/hiring-manager-view/HiringManagerView.tsx
@@ -5,7 +5,6 @@ import {
     ReactNode,
     SetStateAction,
     useCallback,
-    useLayoutEffect,
     useMemo,
     useRef,
     useState,
@@ -36,7 +35,7 @@ import {
     useTCAShareCertificateModal,
 } from '..'
 import { getTCACertificationPath, getUserTCACertificateSsr } from '../../learn.routes'
-import { clearFCCCertificationTitle, hideSiblings } from '../functions'
+import { clearFCCCertificationTitle } from '../functions'
 import { EnvironmentConfig } from '../../../../config'
 
 import { CertificateModal } from './certificate-modal'
@@ -68,7 +67,6 @@ export interface HiringManagerViewProps {
 // eslint-disable-next-line complexity
 const HiringManagerView: FC<HiringManagerViewProps> = (props: HiringManagerViewProps) => {
     const certificateElRef: MutableRefObject<HTMLDivElement | any> = useRef()
-    const wrapElRef: MutableRefObject<HTMLElement | any> = useRef()
 
     const [certPreviewModalIsOpen, setCertPreviewModalIsOpen]: [boolean, Dispatch<SetStateAction<boolean>>]
         = useState<boolean>(false)
@@ -253,19 +251,8 @@ const HiringManagerView: FC<HiringManagerViewProps> = (props: HiringManagerViewP
         )
     }
 
-    useLayoutEffect(() => {
-        const el: HTMLElement = wrapElRef.current
-        if (!el || !props.isModalView) {
-            return
-        }
-
-        hideSiblings(el.parentElement as HTMLElement)
-        hideSiblings(el.parentElement?.parentElement as HTMLElement)
-
-    })
-
     return !!props.certification && !!props.userProfile ? (
-        <div className={props.isModalView ? styles.modalView : ''} ref={wrapElRef}>
+        <div className={props.isModalView ? styles.modalView : ''}>
             {renderHero()}
 
             <ContentLayout
